Refactor imu error and reproj error terms
Previously the t0 and t1 relative timestamps for the imu error are based
on image timestamps. This refactor changes so that the t0 and t1 are
based on the original first and last imu measurement timestamps, instead
of interpolating the imu measurements at t0 and t1 according to image
timestamps (as per OKVIS). This small change allows me to refactor the
reprojection error term to estimate the time-delay according to "Online
Temporal Calibration for Monocular Visual-Inertial Systems" by Qin and
Shen (2018).